+++ deployment/terraform/container_service.tf
@@ -130,8 +130,8 @@ data "template_file" "app" {
     external_domain                  = var.r53_public_hosted_zone
     default_from_email               = var.default_from_email
     notification_email_to            = var.notification_email_to
-    mailchimp_api_key                = var.mailchimp_api_key
-    mailchimp_list_id                = var.mailchimp_list_id
+    hubspot_api_key                  = var.hubspot_api_key
+    hubspot_subscription_id          = var.hubspot_subscription_id
     app_port                         = var.app_port
     aws_region                       = var.aws_region
     project                          = var.project
@@ -178,8 +178,8 @@ data "template_file" "app_cli" {
     external_domain                  = var.r53_public_hosted_zone
     default_from_email               = var.default_from_email
     notification_email_to            = var.notification_email_to
-    mailchimp_api_key                = var.mailchimp_api_key
-    mailchimp_list_id                = var.mailchimp_list_id
+    hubspot_api_key                  = var.hubspot_api_key
+    hubspot_subscription_id          = var.hubspot_subscription_id
     app_port                         = var.app_port
     aws_region                       = var.aws_region
     project                          = var.project